requirements.txt
click==8.0.4
DateTime==4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
SQLAlchemy==1.4.31
uuid==1.30
Werkzeug==2.0.3
zope.interface==5.4.0
